[BUGFIX] Temporary Fix to generatePageTitle failing
If a page has fe_usergroups set $this->getTypoScriptFrontendController()->generatePageTitle() will fail in Backend Context (e.g. you get a 500 when navigating to the given page settings). This is probably an edge case and not really an issue in cs_seo but more so for the core function. Could actually be fixed in TYPO3 13 wasn't really an option.

For now this just falls back to the default title instead of producing a 500 in the backend.